+++ src/main/java/com/gts/backcommons/ssi/dtos/CommonFunctionalityDTO.java
@@ -1,10 +1,12 @@
 package com.gts.backcommons.ssi.dtos;
 
 import lombok.Data;
+import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
 
 @Data
 @SuperBuilder
+@NoArgsConstructor
 public class CommonFunctionalityDTO {
 
     private Long id;