+++ src/main/java/com/gts/backcommons/ssi/dtos/CommonModuleDTO.java
@@ -2,10 +2,12 @@ package com.gts.backcommons.ssi.dtos;
 
 
 import lombok.Data;
+import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
 
 @Data
 @SuperBuilder
+@NoArgsConstructor
 public class CommonModuleDTO {
 
     private Long id;